feat(TDI-40461): refactor only
